splits files out
