+++ src/Encrypt.Config.Json/Encrypt.Config.Json.csproj
@@ -44,7 +44,6 @@
   </ItemGroup>
   <ItemGroup>
     <Compile Include="FileEncrypter.cs" />
-    <Compile Include="JsonConfigurationFileParser.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
   </ItemGroup>
   <ItemGroup>
