--- src/Encrypt.Config.Json/JsonConfigurationFileParser.cs
@@ -1,102 +1,0 @@
-﻿// Copyright (c) .NET Foundation. All rights reserved.
-// Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
-
-using System;
-using System.Collections.Generic;
-using System.Globalization;
-using System.Linq;
-using Newtonsoft.Json.Linq;
-
-namespace Encrypt.Config.Json
-{
-    internal class JsonConfigurationFileParser
-    {
-        private readonly IDictionary<string, string> _data = new SortedDictionary<string, string>(StringComparer.OrdinalIgnoreCase);
-        private readonly Stack<string> _context = new Stack<string>();
-        private string _currentPath;
-
-        public IDictionary<string, string> Parse(JObject jsonConfig)
-        {
-            _data.Clear();
-
-            VisitJObject(jsonConfig);
-
-            return _data;
-        }
-
-        private void VisitJObject(JObject jObject)
-        {
-            foreach (var property in jObject.Properties())
-            {
-                EnterContext(property.Name);
-                VisitProperty(property);
-                ExitContext();
-            }
-        }
-
-        private void VisitProperty(JProperty property)
-        {
-            VisitToken(property.Value);
-        }
-
-        private void VisitToken(JToken token)
-        {
-            switch (token.Type)
-            {
-                case JTokenType.Object:
-                    VisitJObject(token.Value<JObject>());
-                    break;
-
-                case JTokenType.Array:
-                    VisitArray(token.Value<JArray>());
-                    break;
-
-                case JTokenType.Integer:
-                case JTokenType.Float:
-                case JTokenType.String:
-                case JTokenType.Boolean:
-                case JTokenType.Bytes:
-                case JTokenType.Raw:
-                case JTokenType.Null:
-                    VisitPrimitive(token.Value<JValue>());
-                    break;
-
-                default:
-                    throw new FormatException("FormatError Unsupported JSONToken");
-            }
-        }
-
-        private void VisitArray(JArray array)
-        {
-            for (int index = 0; index < array.Count; index++)
-            {
-                EnterContext(index.ToString());
-                VisitToken(array[index]);
-                ExitContext();
-            }
-        }
-
-        private void VisitPrimitive(JValue data)
-        {
-            var key = _currentPath;
-
-            if (_data.ContainsKey(key))
-            {
-                throw new FormatException($"Key Is Duplicated key: {key}");
-            }
-            _data[key] = data.ToString(CultureInfo.InvariantCulture);
-        }
-
-        private void EnterContext(string context)
-        {
-            _context.Push(context);
-            _currentPath = string.Join(".", _context.Reverse());
-        }
-
-        private void ExitContext()
-        {
-            _context.Pop();
-            _currentPath = string.Join(".", _context.Reverse());
-        }
-    }
-}
